requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.13.1
pytz==2021.3
Shapely==1.8.1.post1
sqlparse==0.4.2
